9e28a23e-661e-463a-99f0-ded95b1db172/floor_price.py
import requests


def FloorPrice(url):

    querystring = {"offset":"0","limit":"300"}

    response = requests.request("GET", url, params=querystring)

    response = response.json()

    floorPrice = response["collection"]["stats"]["floor_price"]

    return floorPrice